minor changes v4

+++ src/lib/components/LSPElement.svelte
@@ -4,26 +4,28 @@
   import ToolboxButton from "$lib/components/ToolboxButton.svelte";
 
   export let source: CE;
-  export let destination: CE;
+  export let destination: CE | undefined;
 
   function createPacket() {
-    network.addPacket(source, destination);
+    if (destination) network.addPacket(source, destination);
   }
 
   function deleteDestination() {
-    if (!confirm("Are you sure you want to delete this LSP?")) return;
+    if (!confirm("Are you sure you want to delete this LSP?") || !destination) return;
     source.deleteEntry(destination.address);
     network.notify();
   }
 </script>
 
-<div>
-  <p>{source.address}</p>
-  <p>→</p>
-  <p>{destination.address}</p>
-  <ToolboxButton text={"Play"} callback={createPacket} />
-  <ToolboxButton text={"Delete"} callback={deleteDestination} />
-</div>
+{#if destination}
+  <div>
+    <p>{source.address}</p>
+    <p>→</p>
+    <p>{destination.address}</p>
+    <ToolboxButton text={"Play"} callback={createPacket} />
+    <ToolboxButton text={"Delete"} callback={deleteDestination} />
+  </div>
+{/if}
 
 <style>
   div {
